refactor: consolidated tiles and onPress functions
